% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardCompValues.R
\docType{data}
\name{CompValues}
\alias{CompValues}
\title{Dataset for the standard components values}
\format{
A named list containing the standard values of each component:
\describe{
  \item{"Res"}{the standard values of R1, R2, R3 and R4}
  \item{"ThBeta"}{the standard values of temperature coefficient}
  \item{"ThVal"}{the standard values of nominal thermistor resistances}
}
}
\usage{
data(CompValues)
}
\description{
A list containing the standard values provided for each component.
"Res" corresponds to the values of R1, R2, R3 and R4,
The values in the same index of "ThBeta" and "ThVal" provides nine
types of (beta, Rb) of a thermistor.
}
\examples{
data(EzGP_data)
}
\keyword{dataset}
